split up config
